Moved save/load merged data functions
