+++ openghg_inversions/data/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+
+from .serialise import load_merged_data
